Server/requirements.txt
fastapi
uvicorn
PyPDF2
langchain
langchain_google_genai 
faiss-cpu
python-dotenv
